# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Party Planner API
  description: API pour la gestion d'événements et de fêtes
  version: 1.0.0
  contact:
    name: Party Planner Support
    email: support@party-planner.com
servers:
  - url: ${API_URL}/api  # http://localhost:8000/api
    description: Serveur de développement
  - url: ${API_URL}/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    Photo:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        thumbnail_url:
          type: string
        type:
          $ref: '#/components/schemas/PhotoType'
        caption:
          type: string
          nullable: true
        is_featured:
          type: boolean
        created_at:
          type: string
          format: date-time
      example:
        id: 55
        url: "https://storage.party-planner.com/photos/wedding-001.jpg"
        thumbnail_url: "https://storage.party-planner.com/photos/thumbs/wedding-001.jpg"
        type: "event_photo"
        caption: "La ceremonie en plein air"
        is_featured: true
        created_at: "2024-06-15T16:30:00Z"
...
    PhotoType:
      type: string
      enum: [moodboard, event_photo]
...
paths:
...
  /events/{eventId}/photos:
    get:
      tags: [Photos]
      summary: Lister les photos
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des photos
          content:
            application/json:
              schema:
                type: object
                properties:
                  photos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Photo'
                  stats:
                    type: object

    post:
      tags: [Photos]
      summary: Uploader une photo
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
                type:
                  $ref: '#/components/schemas/PhotoType'
                caption:
                  type: string
      responses:
        '201':
          description: Photo uploadée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
...